+++ apworld/brotato/__init__.py
@@ -4,9 +4,10 @@ from dataclasses import asdict
 from typing import Any, ClassVar, Dict, List, Literal, Set, Tuple, Union
 
 from BaseClasses import Item, LocationProgressType, MultiWorld, Region, Tutorial
-from Options import OptionError
+from Options import OptionError, OptionGroup
 from worlds.AutoWorld import WebWorld, World
 
+from . import options  # So we don't need to import every option class when defining option groups
 from ._loot_crate_groups import BrotatoLootCrateGroup, build_loot_crate_groups
 from .constants import (
     ABYSSAL_TERRORS_CHARACTERS,
@@ -54,6 +55,48 @@ class BrotatoWeb(WebWorld):
     ]
     theme = "dirt"
     rich_text_options_doc = True
+
+    option_groups = [
+        OptionGroup(
+            "Loot Crates",
+            [
+                options.SpawnNormalLootCrates,
+                options.NumberCommonCrateDropLocations,
+                options.NumberCommonCrateDropsPerCheck,
+                options.NumberCommonCrateDropGroups,
+                options.NumberLegendaryCrateDropLocations,
+                options.NumberLegendaryCrateDropsPerCheck,
+                options.NumberLegendaryCrateDropGroups,
+            ],
+        ),
+        OptionGroup(
+            "Item Rewards",
+            [
+                options.ItemWeights,
+                options.CommonItemWeight,
+                options.UncommonItemWeight,
+                options.RareItemWeight,
+                options.LegendaryItemWeight,
+            ],
+        ),
+        OptionGroup(
+            "Upgrades",
+            [
+                options.NumberCommonUpgrades,
+                options.NumberUncommonUpgrades,
+                options.NumberRareUpgrades,
+                options.NumberLegendaryUpgrades,
+            ],
+        ),
+        OptionGroup(
+            "Shop Slots",
+            [options.StartingShopSlots, options.StartingShopLockButtonsMode, options.NumberStartingShopLockButtons],
+        ),
+        OptionGroup(
+            "Abyssal Terrors DLC",
+            [options.EnableAbyssalTerrorsDLC, options.IncludeAbyssalTerrorsCharacters],
+        ),
+    ]
 
 
 class BrotatoWorld(World):
@@ -309,11 +352,19 @@ class BrotatoWorld(World):
         return self.random.choice(self._filler_items)
 
     def fill_slot_data(self) -> Dict[str, Any]:
+        # Define outside dict for readability
+        spawn_normal_loot_crates = (
+            self.options.spawn_normal_loot_crates.value == self.options.spawn_normal_loot_crates.option_true
+        )
         return {
             "waves_with_checks": self.waves_with_checks,
             "num_wins_needed": self.options.num_victories.value,
+            "gold_reward_mode": self.options.gold_reward_mode.value,
+            "xp_reward_mode": self.options.xp_reward_mode.value,
+            "enable_enemy_xp": self.options.enable_enemy_xp.value == self.options.enable_enemy_xp.option_true,
             "num_starting_shop_slots": self.options.num_starting_shop_slots.value,
             "num_starting_shop_lock_buttons": (MAX_SHOP_SLOTS - self.num_shop_lock_button_items),
+            "spawn_normal_loot_crates": spawn_normal_loot_crates,
             "num_common_crate_locations": self.options.num_common_crate_drops.value,
             "num_common_crate_drops_per_check": self.options.num_common_crate_drops_per_check.value,
             "common_crate_drop_groups": [asdict(g) for g in self.common_loot_crate_groups],